Get rid of loopbreak as a mean for signalling successful packet handling.
In blocking mode, breaking the event loop with loopbreak will probably
be needed for exceptions like timeout so we shouldn't use it for
breaking the loop and signalling success.

In the user callbaks used in blocking mode (_rcb and _wcb), disable
read and write events on the event buffer.  This will stop the loop.
Also, set a flag in the packet struct to signal success.

In the "low level callbacks" (_read_cb and _write_cb), enable read or
write, accordingly.