Postpone resolving of DNS names of server.
We used to resolve DNS names when reading configuration.  We now do it
in event_init_socket() and cache the result in the connection object.

The imminent need for changing this is to keep host names around for
X509 certificate verification (CNAME and subjectAltName).  This will
also help later when we implement server failover (and later, when
people want to do more dynamic configuration, f.ex. NAPTR).